requirements.txt
Django
djangorestframework
gunicorn
python-dotenv
django-cors-headers
whitenoise
pillow
pytubefix
requests
